refactor: move the physical-tenant filter to gateway-rest
PhysicalTenantPreSecurityFilter only extracts the physical tenant id from the request URL and stamps
it on PhysicalTenantContext — generic REST-API request plumbing, not a security concern (basic-auth
is one consumer; "pre-security" described its registration order, not its purpose). Rename it to
PhysicalTenantFilter, move it to gateway-rest alongside the routing it pairs with, and register it in
ApiFiltersConfiguration (the central REST-API servlet-filter config) just before Spring Security's
FilterChainProxy. Since the filter no longer needs the configured-tenant set, its registration has
no dependency on authentication beans.

PhysicalTenantSecurityConfiguration now declares only the CamundaSecurityScopeProvider — its genuine
security concern. The shared PhysicalTenantContext stays in spring-utils.
