+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/config/ApiFiltersConfiguration.java
@@ -15,6 +15,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.security.api.model.config.AuthenticationMethod;
 import io.camunda.security.spring.annotation.ConditionalOnAuthenticationMethod;
 import io.camunda.zeebe.gateway.rest.controller.EndpointAccessErrorFilter;
+import io.camunda.zeebe.gateway.rest.controller.PhysicalTenantFilter;
 import io.camunda.zeebe.util.VisibleForTesting;
 import java.util.Arrays;
 import java.util.List;
@@ -46,6 +47,26 @@ public class ApiFiltersConfiguration {
 
   private static List<PathPattern> patterns(final String... patterns) {
     return Arrays.stream(patterns).map(PATTERN_PARSER::parse).toList();
+  }
+
+  // Spring Security's FilterChainProxy registers at SecurityProperties.DEFAULT_FILTER_ORDER (-100);
+  // the physical-tenant filter must run just before it so components inside the security chain can
+  // read the tenant id. We inline the value rather than depend on spring-boot-security for one
+  // constant — and DEFAULT_FILTER_ORDER is a compile-time constant, so a reference would inline the
+  // value anyway and leave the dependency flagged as unused by dependency:analyze. See ADR-0003.
+  private static final int PHYSICAL_TENANT_FILTER_ORDER = -101;
+
+  /**
+   * Stamps the physical tenant id from {@code /physical-tenants/{id}/...} paths onto the request
+   * before Spring Security runs, so the security chain and downstream handlers can read it.
+   */
+  @Bean
+  public FilterRegistrationBean<PhysicalTenantFilter> physicalTenantFilter() {
+    final FilterRegistrationBean<PhysicalTenantFilter> registration =
+        new FilterRegistrationBean<>(new PhysicalTenantFilter());
+    registration.addUrlPatterns(PHYSICAL_TENANTS_PATH_SEGMENT + "*");
+    registration.setOrder(PHYSICAL_TENANT_FILTER_ORDER);
+    return registration;
   }
 
   @ConditionalOnExpression("'${camunda.security.authentication.oidc.groupsClaim:}' != ''")
